requirements.txt
aiohttp==3.9.5
Flask==3.1.0
numpy==2.1.3
openmeteo_requests==1.3.0
pandas==2.2.3
python-dotenv==1.0.1
Requests==2.32.3
requests_cache==1.2.1
retry_requests==2.0.0
scikit_learn==1.4.0
